+++ subsys/bluetooth/audio/Kconfig
@@ -39,4 +39,8 @@ rsource "Kconfig.mpl"
 rsource "Kconfig.mctl"
 rsource "Kconfig.cap"
 
+module = BT_AUDIO
+module-str = "Bluetooth Audio"
+source "subsys/logging/Kconfig.template.log_config"
+
 endif # BT_AUDIO
